renamed SFML folder to UI to prevent naming conflict, added LoadStyles initializer to Source.cpp
