+++ Data/style_button.json
@@ -1,0 +1,14 @@
+{
+  "default": {
+    "bg-color": "0xffffff",
+    "text-color": "0x000000",
+    "border-color": "0xdddddd",
+    "border-size": 0
+  },
+  "hover": {
+    "bg-color": "0xcccccc"
+  },
+  "click": {
+    "bg-color": "0x999999"
+  }
+}
