--- Data/style_menu.json
@@ -1,30 +1,0 @@
-{
-  "title": {
-    "default": {
-      "bg-color": "0xffffff",
-      "text-color": "0x000000",
-      "border-color": "0xdddddd",
-      "border-size": 0
-    },
-    "hover": {
-      "bg-color": "0xcccccc"
-    },
-    "click": {
-      "bg-color": "0x999999"
-    }
-  },
-  "back": {
-    "default": {
-      "bg-color": "0xffffff",
-      "text-color": "0x000000",
-      "border-color": "0xdddddd",
-      "border-size": 0
-    },
-    "hover": {
-      "bg-color": "0xcccccc"
-    },
-    "click": {
-      "bg-color": "0x999999"
-    }
-  }
-}
