+++ SDEV230_Project.vcxproj.filters
@@ -18,36 +18,36 @@
     <None Include="README.md">
       <Filter>Resource Files</Filter>
     </None>
-    <None Include="SFML\SFML.md" />
+    <None Include="UI\SFML.md" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Source.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="SFML\SFML_test.cpp">
+    <ClCompile Include="UI\SFML_test.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="SFML\UserInterface.cpp">
+    <ClCompile Include="UI\UserInterface.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="SFML\UI_Menu.cpp">
+    <ClCompile Include="UI\UI_Menu.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="SFML\Styles.cpp">
+    <ClCompile Include="UI\Styles.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="SFML\SFML_test.h">
+    <ClInclude Include="UI\SFML_test.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="SFML\UserInterface.h">
+    <ClInclude Include="UI\UserInterface.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="SFML\UI_Menu.h">
+    <ClInclude Include="UI\UI_Menu.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="SFML\Styles.h">
+    <ClInclude Include="UI\Styles.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -55,7 +55,7 @@
     <Font Include="Data\arial.ttf" />
   </ItemGroup>
   <ItemGroup>
-    <Text Include="Data\style_product.json" />
-    <Text Include="Data\style_menu.json" />
+    <Text Include="Data\style_background.json" />
+    <Text Include="Data\style_button.json" />
   </ItemGroup>
 </Project>
