+++ Source.cpp
@@ -1,8 +1,17 @@
 #include <iostream>
 
-#include "SFML/SFML_test.h"
-#include "SFML/UserInterface.h"
-#include "SFML/UI_Menu.h"
+#include "UI/SFML_test.h"
+#include "UI/UserInterface.h"
+#include "UI/UI_Menu.h"
+#include "UI/Styles.h"
+
+// Styles.cpp is not implimented yet
+void LoadStyles()
+{
+	using namespace style;
+	LoadStyle("button", "Data/style_button.json");
+	LoadStyle("background", "Data/style_background.json");
+}
 
 int main()
 {
